refactor(pipeline/api-direct-call): extract CarryTypes from IApiDirectCallConfig
Phase 8 commit 3/8 of ApiDirectCallConfig split.

- Carve CarryTypes (IRandomHex16Bootstrap, ISha256Prefix16Bootstrap,
  IJwtClaimBootstrap, SeedCarryBootstrapKind, ISeedCarrySource,
  IDerivedCarry, IWarmStartConfig) from
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/IApiDirectCallConfig.ts.
- New file:
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/ConfigContracts/CarryTypes.ts
  (109 raw LoC). Imports RefToken from ./TemplateTypes.js — bucket-2,
  acyclic per the locked TemplateTypes -> CarryTypes chain.
- IApiDirectCallConfig.ts imports IDerivedCarry, ISeedCarrySource,
  IWarmStartConfig, SeedCarryBootstrapKind and continues to re-export
  them via its existing `export type {}` block so all 53 callers
  compile unchanged.

Verified:
- npm run type-check green
- npx tsup green (lib/index.cjs byte-identical to main)
- npx jest --testPathPatterns=ApiDirectCall green (223/223)
- npx eslint src/Scrapers/Pipeline/Mediator/ApiDirectCall green
- New sub-module 109 LoC (under 150 cap)

Master plan: pipeline-decoupling-master-2026-05-28 / phase-8
